Add CredentialSecret abstract class
Password and Answer are basically the same class. I wanted to use an interface, but
it is better to also have a protected attribute, since again, they're basically the
same class. I kept the tests around because they'll continue to convince me that the
abstract class works as expected. I should probably build a test class that extends
it and test that way, but that is a separate follow-up item.